Fix a run-time integer overflow

+++ vp9/vp9_cx_iface.c
@@ -805,7 +805,7 @@ static int write_superframe_index(vpx_codec_alg_priv_t *ctx) {
 }
 
 // vp9 uses 10,000,000 ticks/second as time stamp
-#define TICKS_PER_SEC 10000000
+#define TICKS_PER_SEC ((int64_t)10000000)
 
 static int64_t timebase_units_to_ticks(const vpx_rational_t *timebase,
                                        int64_t n) {
